PyTorch model's API has been created
