--- kaan/README.txt
@@ -1,60 +1,0 @@
-Image Classification API
--This project provides a RESTful API service for image classification using a pre-trained neural network model with FastAPI.
-
-    1-Features
-        -Predict image class from a supplied image URL.
-        -Utilize a pre-trained Keras model for prediction.
-        -Enable CORS for cross-origin AJAX requests.
-        -Run the service using Docker for easy deployment and isolation.
-
-    2-Requirements
-        -Python 3.11.8 is required. All required libraries are listed in requirements.txt for easy installation.
-
-    3-Installation
-        -To install the required dependencies, run the following command:
-            -pip install -r requirements.txt
-
-    4-Usage
-        -Once the dependencies are installed, you can start the FastAPI server locally using:
-            -uvicorn main.classify:app --host 0.0.0.0 --port 8000
-        -The server will be available at http://127.0.0.1:8000.
-
-    5-Endpoints
-        -GET /: Returns a welcome message.
-        -POST /classify: Accepts a JSON object with an image_url field and returns image classification results.
-        
-    A- Request Format
-        -To classify an image, send a POST request to /classify with a JSON body like the following:
-
-            json
-            {
-              "image_url": "http://example.com/image.jpg"
-            }
-
-    B- Response Format
-        -The API will return a JSON response containing the image classification predictions. For example:
-
-            json
-            {
-              "predictions": [
-                {
-                  "class_name": "Apple",
-                  "probability": "95.47%"
-                },
-                ...
-              ]
-            }
-
-    6-Docker
-        -A Dockerfile is provided to build and run the service in a Docker container.
-
-            -To build the container image, use:
-
-                docker build -t image_classification_api .
-
-            -To run the service in a container, execute:
-
-                docker run -p 8000:8000 image_classification_api
-    7-Note:
-            -Be sure to include any necessary information for setting up and running the application, as well as any configuration details specific to your implementation. 
-            -If there are any additional steps required to train or update the model being used for image classification, include those details as well for complete documentation.
