--- kaan/main/schemas.py
@@ -1,4 +1,0 @@
-from pydantic import BaseModel
-
-class ImageURL(BaseModel): # Defining Image_url schemas
-    image_url: str
