--- orcun/Dockerfile
@@ -1,12 +1,0 @@
-FROM python:3.10
-
-WORKDIR /src
-
-ADD ./ /src
-
-RUN pip install --upgrade pip
-RUN pip install --no-cache-dir -r requirements.txt
-
-EXPOSE 8000 
-
-CMD ["uvicorn","hosting.main:app","--host","0.0.0.0","--port","8000"]
